ForStreamlit/requirements.txt
streamlit==0.82.0
Pillow==8.3.1
